chore: implemented the main component logic
Created a basic component interface where components themselfs are responsible
for rendering their children, but since it is a function I can check if it is
being called and can warn the user if it was not.

Components are also responsible for extracting their own attributes. They can
Check if they have all that they need, and not any extra either.