chore: fixed the broken component registry
It is now exactly how I wanted it. You now have access to the `Component`
class for both extending, and registering. To make a plugin that is
currently possible like this:

```TypeScript
@Component.register
class MyComponent extends Component { ... }
```

I also changed the `MarkUpRenderer` interface to have the exact same way of
registering, and creating.

```TypeScript
@MarkupRenderer.register
class MyComponent implements MarkupRenderer { ... }
```

Might change it from an interface to a class tho as that provides validation
checks using `instance of`, and allows me to inject default behavior if I
change the API later.